+++ pom.xml
@@ -72,6 +72,18 @@
 			<artifactId>commons-compress</artifactId>
 			<version>1.23.0</version>
 		</dependency>
+		<dependency>
+			<groupId>org.mockito</groupId>
+			<artifactId>mockito-core</artifactId>
+			<version>5.0.0</version>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.mockito</groupId>
+			<artifactId>mockito-junit-jupiter</artifactId>
+			<version>5.0.0</version>
+			<scope>test</scope>
+		</dependency>
 
 
 	</dependencies>
