--- src/main/java/com/example/demo/Entity/Senderer.java
@@ -1,35 +1,0 @@
-package com.example.demo.Entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.List;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Setter
-@Getter
-@Entity
-@Table(name = "senderer")
-public class Senderer {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "id")
-    private Long id;
-
-    @Column(name = "fname")
-    private String fname;
-
-    @Column(name = "lname")
-    private String lname;
-
-    @Column(name = "email")
-    private String email;
-
-    @OneToMany(mappedBy = "senderer")
-    private List<Mail> mails;
-}
