--- src/main/java/com/example/demo/config/TokenLoggingFilter.java
@@ -1,4 +1,0 @@
-package com.example.demo.config;
-
-public class TokenLoggingFilter implements  {
-}
