--- src/main/java/com/example/demo/dto/request/CreateSendererRequest.java
@@ -1,9 +1,0 @@
-package com.example.demo.dto.request;
-
-import com.example.demo.service.SendererService;
-import lombok.*;
-
-@Getter
-@Setter
-public class CreateSendererRequest extends BaseSendererRequest {
-}
