+++ src/main/java/com/example/demo/features/mailSystem/dto/MailDTO.java
@@ -1,0 +1,20 @@
+package com.example.demo.features.mailSystem.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class MailDTO {
+    private Long id;
+    private String header;
+    private String body;
+    private Long sender_id;
+    private List<Long> receiver_ids;
+}
