+++ src/main/java/com/example/demo/features/mailSystem/dto/ReceiverDTO.java
@@ -1,0 +1,22 @@
+package com.example.demo.features.mailSystem.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ReceiverDTO {
+    private Long id;
+    private String fname;
+    private String lname;
+    private String email;
+    private String groupName;
+    private List<Long> mail_ids;
+
+}
