+++ src/main/java/com/example/demo/features/mailSystem/dto/request/BaseMailRequest.java
@@ -1,0 +1,15 @@
+package com.example.demo.features.mailSystem.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Setter
+@Getter
+public class BaseMailRequest {
+    private String header;
+    private String body;
+}
