+++ src/main/java/com/example/demo/features/mailSystem/dto/request/BaseReceiverRequest.java
@@ -1,0 +1,17 @@
+package com.example.demo.features.mailSystem.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Setter
+@Getter
+public class BaseReceiverRequest {
+    private String fname;
+    private String lname;
+    private String email;
+    private String groupName;
+}
