+++ src/main/java/com/example/demo/features/mailSystem/entity/Mail.java
@@ -1,0 +1,41 @@
+package com.example.demo.features.mailSystem.entity;
+
+import com.example.demo.features.user.entity.Senderer;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "mail")
+public class Mail {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name = "header")
+    private String header;
+
+    @Column(name = "body")
+    private String body;
+
+    @ManyToOne(optional = true)
+    @JoinColumn(name = "sender_id")
+    private Senderer senderer;
+
+    @ManyToMany
+    @JoinTable(
+            name = "mail_receiver",
+            joinColumns = @JoinColumn(name = "mail_id"),
+            inverseJoinColumns = @JoinColumn(name = "receiver_id")
+    )
+    private List<Receiver> receivers;
+}
