+++ src/main/java/com/example/demo/features/mailSystem/entity/Receiver.java
@@ -1,0 +1,38 @@
+package com.example.demo.features.mailSystem.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.List;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Setter
+@Getter
+@Entity
+@Table(name = "receiver")
+public class Receiver {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name = "fname")
+    private String fname;
+
+    @Column(name = "lname")
+    private String lname;
+
+    @Column(name = "email")
+    private String email;
+
+    @Column(name = "group_name")
+    private String groupName;
+
+    @ManyToMany(mappedBy = "receivers")
+    private List<Mail> mails;
+
+}
