+++ src/main/java/com/example/demo/features/mailSystem/repository/MailRepository.java
@@ -1,0 +1,13 @@
+package com.example.demo.features.mailSystem.repository;
+
+import com.example.demo.features.mailSystem.entity.Mail;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MailRepository extends JpaRepository<Mail,Long> {
+    @Query("SELECT m FROM Mail m WHERE m.id = ?1")
+    public Mail findMailById(Long id);
+
+}
