+++ src/main/java/com/example/demo/features/mailSystem/service/ExcelReaderService.java
@@ -1,0 +1,13 @@
+package com.example.demo.features.mailSystem.service;
+
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+
+public interface ExcelReaderService {
+    public void readRecipientsFromExcel(MultipartFile file) throws IOException;
+
+    }
+
+
+
