+++ src/main/java/com/example/demo/features/user/dto/request/CreateSendererRequest.java
@@ -1,0 +1,8 @@
+package com.example.demo.features.user.dto.request;
+
+import lombok.*;
+
+@Getter
+@Setter
+public class CreateSendererRequest extends BaseSendererRequest {
+}
