+++ src/main/java/com/example/demo/features/user/entity/Senderer.java
@@ -1,0 +1,36 @@
+package com.example.demo.features.user.entity;
+
+import com.example.demo.features.mailSystem.entity.Mail;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.List;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Setter
+@Getter
+@Entity
+@Table(name = "senderer")
+public class Senderer {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name = "fname")
+    private String fname;
+
+    @Column(name = "lname")
+    private String lname;
+
+    @Column(name = "email")
+    private String email;
+
+    @OneToMany(mappedBy = "senderer")
+    private List<Mail> mails;
+}
