--- src/main/java/com/example/demo/repo/MailRepository.java
@@ -1,11 +1,0 @@
-package com.example.demo.repo;
-
-import com.example.demo.Entity.Mail;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface MailRepository extends JpaRepository<Mail,Long> {
-}
