--- src/main/java/com/example/demo/repo/ReceiverRepository.java
@@ -1,9 +1,0 @@
-package com.example.demo.repo;
-
-import com.example.demo.Entity.Receiver;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ReceiverRepository extends JpaRepository<Receiver,Long> {
-}
