--- src/main/java/com/example/demo/service/ExcelReaderServiceImpl.java
@@ -1,55 +1,0 @@
-package com.example.demo.service;
-
-import com.example.demo.Entity.Receiver;
-import com.example.demo.dto.ReceiverDTO;
-import com.example.demo.dto.request.CreateMailRequest;
-import com.example.demo.repo.ReceiverRepository;
-import org.apache.poi.ss.usermodel.Row;
-import org.apache.poi.ss.usermodel.Sheet;
-import org.apache.poi.ss.usermodel.Workbook;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-
-@Service
-public class ExcelReaderServiceImpl implements ExcelReaderService{
-    private final ReceiverRepository receiverRepository;
-
-    public ExcelReaderServiceImpl(ReceiverRepository receiverRepository) {
-        this.receiverRepository = receiverRepository;
-    }
-
-    @Override
-    public void readRecipientsFromExcel(MultipartFile file) throws IOException {
-        List<ReceiverDTO> recipients = new ArrayList<>();
-
-        try (Workbook workbook = new XSSFWorkbook(file.getInputStream())) {
-            Sheet sheet = workbook.getSheetAt(0);  // İlk sayfa
-            for (Row row : sheet) {
-                if (row.getRowNum() == 0) continue; // Başlık satırını atla
-
-                // Hücrelerden verileri oku
-                String name = row.getCell(0).getStringCellValue();
-                String surname = row.getCell(1).getStringCellValue();
-                String email = row.getCell(2).getStringCellValue();
-                String groupName = row.getCell(3).getStringCellValue();
-
-                // Convert DTO to entity and save
-                Receiver receiver = new Receiver();
-                receiver.setFname(name);
-                receiver.setLname(surname);
-                receiver.setEmail(email);
-                receiver.setGroupName(groupName);
-                receiverRepository.save(receiver);
-            }
-        }
-
-
-    }
-}
